+++ History.md
@@ -1,3 +1,9 @@
+0.1.1 / TDB
+==================
+* Adding remaining Blob functionality:
+	- Retrieve and update blob metadata + properties
+	- Delete blobs
+
 0.1.0 / 2011-07-31 
 ==================
 * Initial Release: Basic support for Blobs